fix: 누락된 API 추가 및 admin, user, guest 패키지 분리
