--- src/main/java/com/tookscan/tookscan/order/application/usecase/CreateOrderMemoUseCase.java
@@ -1,9 +1,0 @@
-package com.tookscan.tookscan.order.application.usecase;
-
-import com.tookscan.tookscan.core.annotation.bean.UseCase;
-import com.tookscan.tookscan.order.application.dto.request.CreateOrderMemoRequestDto;
-
-@UseCase
-public interface CreateOrderMemoUseCase {
-    void execute(Long orderId, CreateOrderMemoRequestDto requestDto);
-}
