--- src/main/java/com/tookscan/tookscan/order/application/usecase/ReadUserOrderDetailUseCase.java
@@ -1,10 +1,0 @@
-package com.tookscan.tookscan.order.application.usecase;
-
-import com.tookscan.tookscan.core.annotation.bean.UseCase;
-import com.tookscan.tookscan.order.application.dto.response.ReadUserOrderDetailResponseDto;
-import java.util.UUID;
-
-@UseCase
-public interface ReadUserOrderDetailUseCase {
-    ReadUserOrderDetailResponseDto execute(UUID accountID, Long orderId);
-}
