+++ src/main/java/com/tookscan/tookscan/order/application/user/usecase/UpdateOrderScanUseCase.java
@@ -1,0 +1,9 @@
+package com.tookscan.tookscan.order.application.user.usecase;
+
+import com.tookscan.tookscan.core.annotation.bean.UseCase;
+import java.util.UUID;
+
+@UseCase
+public interface UpdateOrderScanUseCase {
+    void execute(UUID accountId, Long orderId);
+}
